+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] - 2019-06-17
+* Added support for changing the header on every page now
+* Changed the extension-config
+
 ## [2.0.0] - 2019-01-15
 * Updates for extension to work with PWA 6.1. Removed usage of subscription stream and instead update based on current pathname.
 
@@ -13,6 +17,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [1.0.0] - 2018-11-28
 - first stable release
 
+[2.0.0]: https://github.com/shopgate/ext-change-header-colors/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/shopgate/ext-change-header-colors/compare/v1.0.1...v2.0.0
 [1.0.1]: https://github.com/shopgate/ext-change-header-colors/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/shopgate/ext-change-header-colors/tree/v1.0.0